+++ CHANGELOG.md
@@ -49,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - updated upload and download artifacts actions to v4
 
 ### Fixed
+- do not break long file content lines
 - source TF_WORKING_DIR from env helper instead of process.env in locals helper
 - fixed how terraform state is accessed before it the initial synchronization
 - links to supported resources in HOWTOs